Moved H2 DB location to temp directory
The H2 inter-process communication DB was created in the user.home but Java 1.7 and lower can have a different opinion than java 1.8 about what the user.home directory is. With the system temp directory there should be no ambiguity and it's a better place for this kind of DB anyway.