Switch to slf4j for logging
 - Replace direct log4j calls with slf4j.
 - In taf-base move log4j to test scope and log4j xml configuration
   file into test/resources.
 - In taf-swing-server leave log4j and its configuration files in
   compile scope since it's effectively an end-user application.

This removes the imposition of a specific logging implementation on
our end-user projects which are then free to choose their preferred
logging implementation and configuration without conflicts.

This means projects which before this change relied on TAF logging
implementation and configuration will need to initialize logging
correctly on their side.

Good in-depth explanation of the rationale: http://veithen.github.io/2009/10/16/maven-manage-slf4j-deps.html