Add module to build the Java 8 version of taf-swing-server
The taf-swing-server component runs in the JVM of the application under
test. If this JVM is version 8, the assertj-swing-junit dependency
needs to be 3.x, otherwise the tests will run unreliably.

assertj-swing-junit 3.x requires at least Java 8, so simply upgrading
the dependency version is not feasible as long as testing Java 6
applications is required.

The new taf-swing-server-java8 module takes the Java 6 taf-swing-server
source code, overrides the assertj-swing-junit dependency with version
3.x, then compiles and assembles the required JARs.