Remove JUnit compile dependency
TAF required a compile dependency to JUnit. This dependency was
required to support asserts and extensions to JUnit 4 Rules, but as
consequence imposed JUnit4 as transitive dependency to client projects.

 - JUnit Asserts in main code have been reimplemented.
 - Extensions to JUnit 4 have been isolated in new, dedicated modules.

This allows the "common" modules to avoid the compile dependency to
JUnit, meaning that client projects don't have it imposed to them
transitively anymore.

Client projects still relying on JUnit 4 extensions need to import the
required additional ad-hoc dependencies. Equivalent JUnit 5 extensions
are not yet implemented.