Make taf-browser resilient to stale elements
Pages with Javascript modifying the DOM can lead to elements being
removed from the DOM between the time they are found and an operation
is attempted on them. This can lead to StaleElementReferenceException
exceptions being encountered and breaking the execution.

This change implements safe operation wrappers which re-try the
operation until either successful or a timeout elapses.

All taf-browser Br-elements have been upgraded to use the new safe
operations .