Implement unit test for stale element race condition
Stale elements are the result of a race condition where the element is
removed from the DOM between the WebElement instance being acquired and
an operation being performed on it.

To reliably test such a race condition, a mock button class is
implemented with the test instance as parameter. The reference to the
test instance is used to trigger the button removal in the page DOM at
the appropriate time while the click operation is being performed.