Migrate taf-swing-server to JUnit 5
Previously taf-swing-server had a dependency on assertj-swing-junit,
which included a transitive dependency to JUnit 4.

This change replaces said dependecy with assertj-swing, which still
provides the fixtures required by taf-swing-server to operate, but
doesn't include JUnit 4.

JUnit 5 is added as explicit dependency in test scope and the tests have
been migrated accordingly.